Close connection at the end
Huawei routers support limitted number of concurrent sessions. If the pool is
already depleted a new session will not be created even when valid credentials
are passed. If scrapping is performed frequently enough unclosed sessions from
previous scrapes can be still around when a newly commenced scrape will fail
during session setup.  To prevent that problem session needs to be closed and
returned to the free pool after scrape is finished.